Add flag to control k8s client set creation
Now bool flag can be supplied via CLI that toggles creation of k8s
client set for agent checker. It is needed for the server's test runs
when there is no need of kubernetes API querying.

Docker file for build container image is cleaned from getting the project's
dependencies inside the container. The whole vendor directory will be
synced when starting the container.