CollectAgentsMetrics logic depends on persistent storage in use
Previously used method of checking outdated information about agents
does not work properly when etcd is in use (it reports false positive
when data from agents is not updated). Now, dedicated branch is added
for the case of using etcd and Agents.CheckAgents() is called there
to get agents whose data is outdated (missing).
Also, interval of checking that agents data is up-to-date can now be
set via command-line ("-check-interval") in seconds.