Increase delay in e2e to test agent updates
With 15 sec delay we're testing that server is able to add agents
to K8s Custom Resources. But we're not checking that server is
able to update them and make sure agents are up to date.

Increasing delay to 45 seconds will make sure server/agent had
enough time to detect outdated state.